feat: Payment 및 PaymentDiscount 적용

+++ src/main/java/flight_booking/demo/domain/discount/repository/DiscountRepository.java
@@ -3,5 +3,5 @@ package flight_booking.demo.domain.discount.repository;
 import flight_booking.demo.domain.discount.entity.Discount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DiscountRepository extends JpaRepository<Discount, Integer> {
+public interface DiscountRepository extends JpaRepository<Discount, Long> {
 }
